refactor: extract bash shell syntax
