+++ crates/tools/src/lib.rs
@@ -8,6 +8,7 @@
 //! - Read-only exploration: read_file, grep, glob, list_dir
 //! - All: core + read-only exploration tools
 
+mod bash_classifier;
 mod exploration_tools;
 mod file_tools;
 
@@ -20,6 +21,7 @@ use exploration_tools::is_binary_file;
 use file_tools::{detect_mime, is_image};
 
 use alan_agent_engine::tools::{Tool, ToolContext, ToolRegistry, ToolResult};
+use bash_classifier::classify_bash_command;
 use serde_json::{Value, json};
 use std::path::Path;
 
@@ -34,264 +36,6 @@ pub struct BashTool;
 impl BashTool {
     pub fn new() -> Self {
         Self
-    }
-}
-
-fn classify_bash_command(command: &str) -> alan_agent_protocol::ToolCapability {
-    let normalized = normalize_shell_line_continuations(command).to_lowercase();
-    let fragments = split_shell_fragments(&normalized);
-
-    let mut saw_write = false;
-    let mut saw_unknown = false;
-    for fragment in fragments {
-        let capability = classify_bash_fragment(fragment.trim());
-        if matches!(capability, alan_agent_protocol::ToolCapability::Network) {
-            return alan_agent_protocol::ToolCapability::Network;
-        }
-        if matches!(capability, alan_agent_protocol::ToolCapability::Write) {
-            saw_write = true;
-        }
-        if matches!(capability, alan_agent_protocol::ToolCapability::Unknown) {
-            saw_unknown = true;
-        }
-    }
-
-    if saw_write {
-        alan_agent_protocol::ToolCapability::Write
-    } else if saw_unknown {
-        alan_agent_protocol::ToolCapability::Unknown
-    } else {
-        alan_agent_protocol::ToolCapability::Read
-    }
-}
-
-fn is_shell_word_boundary(ch: char) -> bool {
-    ch.is_whitespace() || matches!(ch, ';' | '|' | '&' | '(' | ')' | '<' | '>' | '{' | '}')
-}
-
-fn normalize_shell_line_continuations(command: &str) -> String {
-    let mut normalized = String::with_capacity(command.len());
-    let mut chars = command.chars().peekable();
-    let mut in_single = false;
-    let mut in_double = false;
-    let mut in_comment = false;
-    let mut escaped = false;
-    let mut word_started = false;
-
-    while let Some(ch) = chars.next() {
-        if in_comment {
-            normalized.push(ch);
-            if matches!(ch, '\n' | '\r') {
-                in_comment = false;
-                word_started = false;
-            }
-            continue;
-        }
-
-        if escaped {
-            normalized.push(ch);
-            escaped = false;
-            word_started = true;
-            continue;
-        }
-
-        if in_single {
-            if ch == '\'' {
-                in_single = false;
-            }
-            normalized.push(ch);
-            word_started = true;
-            continue;
-        }
-
-        if in_double {
-            match ch {
-                '\\' => {
-                    if consume_shell_line_continuation(&mut chars) {
-                        continue;
-                    }
-                    normalized.push(ch);
-                    escaped = true;
-                }
-                '"' => {
-                    in_double = false;
-                    normalized.push(ch);
-                    word_started = true;
-                }
-                _ => {
-                    normalized.push(ch);
-                    word_started = true;
-                }
-            }
-            continue;
-        }
-
-        match ch {
-            '\\' => {
-                if consume_shell_line_continuation(&mut chars) {
-                    continue;
-                }
-                normalized.push(ch);
-                escaped = true;
-                word_started = true;
-            }
-            '\'' => {
-                in_single = true;
-                normalized.push(ch);
-                word_started = true;
-            }
-            '"' => {
-                in_double = true;
-                normalized.push(ch);
-                word_started = true;
-            }
-            '#' if !word_started => {
-                in_comment = true;
-                normalized.push(ch);
-            }
-            c if is_shell_word_boundary(c) => {
-                normalized.push(c);
-                word_started = false;
-            }
-            _ => {
-                normalized.push(ch);
-                word_started = true;
-            }
-        }
-    }
-
-    normalized
-}
-
-fn split_shell_fragments(command: &str) -> Vec<String> {
-    let mut fragments = Vec::new();
-    let mut current = String::with_capacity(command.len());
-    let mut chars = command.chars().peekable();
-    let mut in_single = false;
-    let mut in_double = false;
-    let mut in_comment = false;
-    let mut escaped = false;
-    let mut word_started = false;
-
-    while let Some(ch) = chars.next() {
-        if in_comment {
-            if matches!(ch, '\n' | '\r') {
-                push_shell_fragment(&mut fragments, &mut current);
-                in_comment = false;
-                word_started = false;
-            }
-            continue;
-        }
-
-        if escaped {
-            current.push(ch);
-            escaped = false;
-            word_started = true;
-            continue;
-        }
-
-        if in_single {
-            if ch == '\'' {
-                in_single = false;
-            }
-            current.push(ch);
-            word_started = true;
-            continue;
-        }
-
-        if in_double {
-            match ch {
-                '\\' => {
-                    current.push(ch);
-                    escaped = true;
-                }
-                '"' => {
-                    in_double = false;
-                    current.push(ch);
-                    word_started = true;
-                }
-                _ => {
-                    current.push(ch);
-                    word_started = true;
-                }
-            }
-            continue;
-        }
-
-        match ch {
-            '\\' => {
-                current.push(ch);
-                escaped = true;
-                word_started = true;
-            }
-            '\'' => {
-                in_single = true;
-                current.push(ch);
-                word_started = true;
-            }
-            '"' => {
-                in_double = true;
-                current.push(ch);
-                word_started = true;
-            }
-            '#' if !word_started => {
-                in_comment = true;
-            }
-            '&' if matches!(chars.peek(), Some('&')) => {
-                chars.next();
-                push_shell_fragment(&mut fragments, &mut current);
-                word_started = false;
-            }
-            '|' if matches!(chars.peek(), Some('|')) => {
-                chars.next();
-                push_shell_fragment(&mut fragments, &mut current);
-                word_started = false;
-            }
-            ';' | '\n' | '\r' | '|' => {
-                push_shell_fragment(&mut fragments, &mut current);
-                word_started = false;
-            }
-            c if is_shell_word_boundary(c) => {
-                current.push(c);
-                word_started = false;
-            }
-            _ => {
-                current.push(ch);
-                word_started = true;
-            }
-        }
-    }
-
-    push_shell_fragment(&mut fragments, &mut current);
-    fragments
-}
-
-fn push_shell_fragment(fragments: &mut Vec<String>, current: &mut String) {
-    if current.trim().is_empty() {
-        current.clear();
-        return;
-    }
-
-    fragments.push(std::mem::take(current));
-}
-
-fn consume_shell_line_continuation<I>(chars: &mut std::iter::Peekable<I>) -> bool
-where
-    I: Iterator<Item = char>,
-{
-    match chars.peek().copied() {
-        Some('\n') => {
-            chars.next();
-            true
-        }
-        Some('\r') => {
-            chars.next();
-            if matches!(chars.peek(), Some('\n')) {
-                chars.next();
-            }
-            true
-        }
-        _ => false,
     }
 }
